requirements.txt
fastapi
uvicorn
requests
transformers
torch
gradio
pydantic
python-dotenv
huggingface_hub
